requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
twilio==8.2.2
PyJWT==2.7.0
Werkzeug==2.3.4
